ChatGPT/Chatbot/requirements.txt
python-dotenv
openai
langchain
yt_dlp
pydub
pypdf
bs4
tiktoken
langchain[docarray]
pydantic==1.10.9
flask